+++ spotbugsTestCases/src/java/exposemutable/FindReturnRefNegativeTest.java
@@ -183,6 +183,16 @@ public class FindReturnRefNegativeTest {
         }
     }
 
+	public static class PreferenceTreeNode {
+		private PreferenceTreeNode fParent; 
+		public void addChild(PreferenceTreeNode node) {
+			node.fParent= this; 
+		}
+		public PreferenceTreeNode getfParent() {
+			return fParent;
+		}
+	} 
+    
     private CharBuffer charBuf;
     private char[] charArray;